Computes the annotation vector that favors complexity

Computes the annotation vector that favors complexity

av_complexity(.mp, data, dilution_factor = 0, apply = FALSE)

Arguments

.mp

a Matrix Profile object.

data

a vector or a column matrix of numeric.

dilution_factor

a numeric. (Default is 0). Larger numbers means more dilution.

apply

logical. (Default is FALSE). Applies the Annotation Vector over the Matrix Profile. Use with caution.

Value

Returns the input .mp object with an embedded annotation vector.

References

Dau HA, Keogh E. Matrix Profile V: A Generic Technique to Incorporate Domain Knowledge into Motif Discovery. In: Proceedings of the 23rd ACM SIGKDD International Conference on Knowledge Discovery and Data Mining - KDD '17. New York, New York, USA: ACM Press; 2017. p. 125-34.

See also

Examples

data <- mp_test_data$train$data[1:1000]
w <- 50
mp <- tsmp(data, window_size = w, verbose = 0)
av <- av_complexity(mp, apply = TRUE)