FLOSS - Extract Segments

Extract candidate points of semantic changes.

floss_extract(.mpac, threshold = 1, exclusion_zone = NULL)

Arguments

.mpac

a TSMP object of class ArcCount.

threshold

a number. (Default is 1). Set the maximum value for evaluating semantic changes. This is data specific. It is advised to check what is 'normal' for your data.

exclusion_zone

if a number will be used instead of embedded value. (Default is NULL).

Value

Returns the input .mp object a new name floss with the location of semantic changes and floss_vals with the normalized arc count value of the semantic change positions.

References

Gharghabi S, Ding Y, Yeh C-CM, Kamgar K, Ulanova L, Keogh E. Matrix Profile VIII: Domain Agnostic Online Semantic Segmentation at Superhuman Performance Levels. In: 2017 IEEE International Conference on Data Mining (ICDM). IEEE; 2017. p. 117-26.

Website: https://sites.google.com/site/onlinesemanticsegmentation/

Website: http://www.cs.ucr.edu/~eamonn/MatrixProfile.html

See also

Examples

data <- mp_fluss_data$tilt_abp$data[1:1000]
w <- 10
mp <- tsmp(data, window_size = w, verbose = 0)
mp <- fluss_cac(mp)
mp <- fluss_extract(mp, 2)