FLUSS - Prediction score calculation

FLUSS - Prediction score calculation

fluss.score(gtruth, extracted, data.size)

Arguments

gtruth
an int or vector of int with the ground truth index of segments.

extracted
an int or vector of int with the extracted indexes from fluss.extract().

data.size
an int. Size of original input data.

Value

Returns the score of predicted semantic transitions compared with the ground truth. Zero is the best, One is the worst.

References

Gharghabi S, Ding Y, Yeh C-CM, Kamgar K, Ulanova L, Keogh E. Matrix Profile VIII: Domain Agnostic Online Semantic Segmentation at Superhuman Performance Levels. In: 2017 IEEE International Conference on Data Mining (ICDM). IEEE; 2017. p. 117–26.

Website: https://sites.google.com/site/onlinesemanticsegmentation/

Website: http://www.cs.ucr.edu/~eamonn/MatrixProfile.html

See also

Examples

# This is a fast toy example and results are useless. For a complete result, run the code inside
#'Not run' section below.
data <- fluss_data$tilt.abp$data[1:1000]
w <- 10
truth <- 400
mp <- mstomp(data, w, verbose = 0)
cac <- fluss.cac(mp$pi, w)
segments <- fluss.extract(cac, 1, w)
score <- fluss.score(truth, segments, length(data))
# NOT RUN {
data <- fluss_data$walkjogrun$data
w <- fluss_data$walkjogrun$window # 80
truth <- fluss_data$walkjogrun$gtruth # 3800 6800
nseg <- length(fluss_data$walkjogrun$gtruth) # 2
mp <- mstomp(data, w)
cac <- fluss.cac(mp$pi, w)
segments <- fluss.extract(cac, nseg, w)
score <- fluss.score(truth, segments, length(data))
# }