Computes the golden section for combined candidates

Computes the golden section for combined candidates

golden.section.2(dist.pro, thold, label, pos.st, pos.ed, beta, window.size,
  fit.idx)

Arguments

dist.pro
the candidate distance profile

thold
a number with the threshold used to calculate the F-Score

label
a vector with the data bool annotation

pos.st
a vector with the starting points of label

pos.ed
a vector with the ending points of label

beta
a number that balance the F-Score. Beta > 1 towards recall, < towards precision

window.size
an integer with the sliding window size

fit.idx
an integer with the index of the current threshold

Value

Returns the best threshold and its F-Score