Calculates the distance profile using MASS_V2 algorithm

Mueen's Algorithm for Similarity Search is The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance and Correlation Coefficient.

mass(data_fft, query_window, data_size, window_size, data_mean, data_sd,
 query_mean, query_sd)

Arguments

data_fft

precomputed data product.

query_window

a vector of numeric. Query window.

data_size

an int. The length of the reference data.

window_size

an int. Sliding window size.

data_mean

precomputed data moving average.

data_sd

precomputed data moving standard deviation.

query_mean

precomputed query average.

query_sd

precomputed query standard deviation.

Value

Returns the distance_profile for the given query and the last_product for STOMP algorithm.

References

Abdullah Mueen, Yan Zhu, Michael Yeh, Kaveh Kamgar, Krishnamurthy Viswanathan, Chetan Kumar Gupta and Eamonn Keogh (2015), The Fastest Similarity Search Algorithm for Time Series Subsequences under Euclidean Distance

Website: https://www.cs.unm.edu/~mueen/FastestSimilaritySearch.html

See also