Multivariate STOMP algorithm Parallel version

Computes the Matrix Profile and Profile Index for Multivariate Time Series.

mstomp.par(data, window.size, must.dim = NULL, exc.dim = NULL,
  exclusion.zone = 1/2, verbose = 2, n.workers = 2)

Arguments

data
a matrix of numeric, where each colums is a time series. Accepts vector (see details), list and data.frame too.

window.size
an int. Size of the sliding window.

must.dim
an int or vector of which dimensions to forcibly include (default is NULL).

exc.dim
an int or vector of which dimensions to exclude (default is NULL).

exclusion.zone
a numeric. Size of the exclusion zone, based on query size (default is 1/2). See details.

verbose
an int. See details. (Default is 2).

n.workers
an int. Number of workers for parallel. (Default is 2).

Value

Returns the matrix profile mp and profile index pi. It also returns the left and right matrix profile lmp, rmp and profile index lpi, rpi that may be used to detect Time Series Chains (Yan Zhu 2018).

Details

The Matrix Profile, has the potential to revolutionize time series data mining because of its generality, versatility, simplicity and scalability. In particular it has implications for time series motif discovery, time series joins, shapelet discovery (classification), density estimation, semantic segmentation, visualization, rule discovery, clustering etc. The MSTOMP computes the Matrix Profile and Profile Index for Multivariate Time Series that is meaningful for multidimensional MOTIF discovery. It uses the STOMP algorithm that is faster than STAMP but lacks its anytime property.

Although this functions handles Multivariate Time Series, it can also be used to handle Univariate Time Series. verbose changes how much information is printed by this function; 0 means nothing, 1 means text, 2 means text and sound.

References

Yeh CM, Kavantzas N, Keogh E. Matrix Profile VI : Meaningful Multidimensional Motif Discovery.

Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1–27.

Website: https://sites.google.com/view/mstamp/

Website: http://www.cs.ucr.edu/~eamonn/MatrixProfile.html

See also

Examples

# using all dimensions
Sys.sleep(1) # sometimes sleep is needed if you run parallel multiple times in a row
mp <- mstomp.par(toy_data$data[1:100,], 30, verbose = 0)