Pan Matrix Profile upper bound

Finds the upper bound for Pan Matrix Profile calculation.

pmp_upper_bound(
  data,
  threshold = getOption("tsmp.pmp_ub", 0.95),
  refine_stepsize = getOption("tsmp.pmp_refine", 0.25),
  return_pmp = TRUE,
  n_workers = 1,
  verbose = getOption("tsmp.verbose", 2)
)

Arguments

data
a matrix or a vector of numeric.

threshold
a numeric. Correlation threshold. See details. (Default is 0.95).

refine_stepsize
a numeric. Step size for the last upper bound search. See details. (Default is 0.25).

return_pmp
a logical. If TRUE, returns the computed data as a PMP object, if FALSE, returns just the upper bound value. (Default is TRUE).

n_workers
an int. Number of workers for parallel. (Default is 1).

verbose
verbose an int. See details. (Default is 2).

Value

Returns a PMP object with computed data, or just the upper bound value if return_pmp is set to FALSE.

Details

The Pan Matrix Profile may not give any further information beyond a certain window size. This function starts computing the matrix profile for the window size of 8 and doubles it until the minimum correlation value found is less than the threshold. After that, it begins to refine the upper bound using the refine_stepsize values, until the threshold value is hit.

verbose changes how much information is printed by this function; 0 means nothing, 1 means text, 2 adds the progress bar, 3 adds the finish sound.

References

Yet to be announced

Website: http://www.cs.ucr.edu/~eamonn/MatrixProfile.html

Examples

# return the object
pan_matrix <- pmp_upper_bound(mp_gait_data)
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf
#> Error in corr_ed(result$mp, window_size): could not find function "corr_ed"

# just the upper bound
pan_ub <- pmp_upper_bound(mp_gait_data, return_pmp = FALSE)
#> Warning: no non-missing arguments to max; returning -Inf
#> Warning: no non-missing arguments to max; returning -Inf