Retrieve salient subsequences from a dataset

In order to allow a meaningful visualization in Multi-Dimensional Space (MDS), this function retrieves the most relevant subsequences using Minimal Description Length (MDL) framework.

salient.subsequences(data, matrix.profile, profile.index, window.size,
  n.bits = 8, n.cand = 10, exclusion.zone = 1/2, verbose = 2)

Arguments

data
a vector, column matrix or data.frame. If more than one column is provided, see details.

matrix.profile
a result from STAMP or STOMP algorithms.

profile.index
a result from STAMP or STOMP algorithms.

window.size
an int with the size of the sliding window.

n.bits
an int. Number of bits for MDL discretization. (Default is 8).

n.cand
an int. number of candidate when picking the subsequence in each iteration. (Default is 10).

exclusion.zone
a numeric. Size of the exclusion zone, based on window.size. (Default is 1/2). See details.

verbose
an int. See details. (Default is 2).

Value

Returns a list with indexes, a vector with the starting position of each subsequence, idx.bit.size, a vector with the associated bitsize for each iteration and bits the value used as input on n.bits.

Details

The main purpose of this algorithm is to find subsequences in one time series, but this implementation also covers the experimental effectiveness evaluation with "whole sequence" setting. This means you can input a matrix where each column is a sequence and this algorithm will retrieve the most relevant sequences. For this setting, the exclusion.zone is ignored, and you need to pre-compute the ordinary euclidean distance matrix. See examples.

The exclusion.zone is used to avoid trivial matches.

verbose changes how much information is printed by this function; 0 means nothing, 1 means text, 2 means text and sound.

References

Yeh CCM, Van Herle H, Keogh E. Matrix profile III: The matrix profile allows visualization of salient subsequences in massive time series. Proc - IEEE Int Conf Data Mining, ICDM. 2017;579–88.

Hu B, Rakthanmanon T, Hao Y, Evans S, Lonardi S, Keogh E. Discovering the Intrinsic Cardinality and Dimensionality of Time Series Using MDL. In: 2011 IEEE 11th International Conference on Data Mining. IEEE; 2011. p. 1086–91.

Website: https://sites.google.com/site/salientsubs/

Examples

# NOT RUN {
  # subsequences setting (main purpose)
  salient.subsequences(data, data$mp, data$pi, 30, n.bits = 8, n.cand = 10)

  # sequences setting
  dist.matrix <- as.matrix(stats::dist(carfull$data))
  mp <- matrix(0, nrow(carfull$data), 1)
  pi <- matrix(0, nrow(carfull$data), 1)

  for (i in 1:nrow(carfull$data)) {
    dist.matrix[i, i] <- Inf;
    pi[i] <- which.min(dist.matrix[i, ])
    mp[i] <- dist.matrix[i, pi[i]]
  }

  n.bits <- 8

  subs <- salient.subsequences(t(carfull$data), mp, pi, 577, n.bits = n.bits, n.cand = 10)
  cutoff <- which(diff(subs$idx.bit.size) > 0)[1] - 1
  if(cutoff > 0) {
    carfull$lab[subs$indexes[1:(cutoff - 1)]]
  } else {
    message("nothing to do")
  }
# }