Framework for retrieve salient subsequences from a dataset

In order to allow a meaningful visualization in Multi-Dimensional Space (MDS), this function retrieves the most relevant subsequences using Minimal Description Length (MDL) framework.

salient_subsequences(
  .mp,
  data,
  n_bits = 8,
  n_cand = 10,
  exclusion_zone = NULL,
  verbose = getOption("tsmp.verbose", 2)
)

Arguments

.mp

a TSMP object of class MatrixProfile.

data

the data used to build the Matrix Profile, if not embedded.

n_bits

an int or vector of int. Number of bits for MDL discretization. (Default is 8).

n_cand

an int. number of candidate when picking the subsequence in each iteration. (Default is 10).

exclusion_zone

if a number will be used instead of embedded value. (Default is NULL).

verbose

an int. See details. (Default is 2).

Value

Returns the input .mp object with a new name salient. It contains: indexes, a vector

with the starting position of each subsequence, idx_bit_size, a vector with the associated bitsize for each iteration and bits the value used as input on n_bits.

Details

verbose changes how much information is printed by this function; 0 means nothing, 1 means text, 2 adds the progress bar, 3 adds the finish sound.

References

Yeh CCM, Van Herle H, Keogh E. Matrix profile III: The matrix profile allows visualization of salient subsequences in massive time series. Proc - IEEE Int Conf Data Mining, ICDM. 2017;579-88.

Hu B, Rakthanmanon T, Hao Y, Evans S, Lonardi S, Keogh E. Discovering the Intrinsic Cardinality and Dimensionality of Time Series Using MDL. In: 2011 IEEE 11th International Conference on Data Mining. IEEE; 2011. p. 1086-91.

Website: https://sites.google.com/site/salientsubs/

Examples

# toy example
data <- mp_toy_data$data[, 1]
mp <- tsmp(data, window_size = 30, verbose = 0)
mps <- salient_subsequences(mp, data, verbose = 0)
# \donttest{
# full example
data <- mp_meat_data$sub$data
w <- mp_meat_data$sub$sub_len
mp <- tsmp(data, window_size = w, verbose = 2, n_workers = 2)
#> Warming up parallel with 2 cores.
#> 
#> STOMP [------------------------------]   1% at 3757 it/s, elapsed:  0s, eta: 28s
#> 
#> STOMP [------------------------------]   1% at 3469 it/s, elapsed:  0s, eta: 30s
#> 
#> STOMP [------------------------------]   2% at 3245 it/s, elapsed:  0s, eta: 32s
#> 
#> STOMP [>-----------------------------]   2% at 3157 it/s, elapsed:  1s, eta: 33s
#> 
#> STOMP [>-----------------------------]   2% at 3082 it/s, elapsed:  1s, eta: 34s
#> 
#> STOMP [>-----------------------------]   2% at 3018 it/s, elapsed:  1s, eta: 35s
#> 
#> STOMP [>-----------------------------]   3% at 2991 it/s, elapsed:  1s, eta: 35s
#> 
#> STOMP [>-----------------------------]   3% at 2980 it/s, elapsed:  1s, eta: 35s
#> 
#> STOMP [>-----------------------------]   3% at 2958 it/s, elapsed:  1s, eta: 35s
#> 
#> STOMP [>-----------------------------]   3% at 2858 it/s, elapsed:  1s, eta: 36s
#> 
#> STOMP [>-----------------------------]   4% at 2794 it/s, elapsed:  1s, eta: 37s
#> 
#> STOMP [>-----------------------------]   4% at 2755 it/s, elapsed:  1s, eta: 37s
#> 
#> STOMP [>-----------------------------]   4% at 2717 it/s, elapsed:  2s, eta: 38s
#> 
#> STOMP [>-----------------------------]   4% at 2654 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [>-----------------------------]   5% at 2632 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [>-----------------------------]   5% at 2617 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [=>----------------------------]   5% at 2604 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [=>----------------------------]   5% at 2576 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [=>----------------------------]   6% at 2569 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [=>----------------------------]   6% at 2577 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [=>----------------------------]   6% at 2588 it/s, elapsed:  2s, eta: 39s
#> 
#> STOMP [=>----------------------------]   6% at 2592 it/s, elapsed:  3s, eta: 39s
#> 
#> STOMP [=>----------------------------]   7% at 2597 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [=>----------------------------]   7% at 2602 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [=>----------------------------]   7% at 2601 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [=>----------------------------]   7% at 2604 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [=>----------------------------]   8% at 2608 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [=>----------------------------]   8% at 2609 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [=>----------------------------]   8% at 2602 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [=>----------------------------]   8% at 2598 it/s, elapsed:  3s, eta: 38s
#> 
#> STOMP [==>---------------------------]   9% at 2603 it/s, elapsed:  4s, eta: 38s
#> 
#> STOMP [==>---------------------------]   9% at 2603 it/s, elapsed:  4s, eta: 37s
#> 
#> STOMP [==>---------------------------]   9% at 2607 it/s, elapsed:  4s, eta: 37s
#> 
#> STOMP [==>---------------------------]   9% at 2612 it/s, elapsed:  4s, eta: 37s
#> 
#> STOMP [==>---------------------------]  10% at 2616 it/s, elapsed:  4s, eta: 37s
#> 
#> STOMP [==>---------------------------]  10% at 2622 it/s, elapsed:  4s, eta: 37s
#> 
#> STOMP [==>---------------------------]  10% at 2628 it/s, elapsed:  4s, eta: 37s
#> 
#> STOMP [==>---------------------------]  10% at 2637 it/s, elapsed:  4s, eta: 36s
#> 
#> STOMP [==>---------------------------]  11% at 2638 it/s, elapsed:  4s, eta: 36s
#> 
#> STOMP [==>---------------------------]  11% at 2642 it/s, elapsed:  4s, eta: 36s
#> 
#> STOMP [==>---------------------------]  11% at 2648 it/s, elapsed:  4s, eta: 36s
#> 
#> STOMP [==>---------------------------]  11% at 2650 it/s, elapsed:  5s, eta: 36s
#> 
#> STOMP [==>---------------------------]  12% at 2654 it/s, elapsed:  5s, eta: 36s
#> 
#> STOMP [===>--------------------------]  12% at 2657 it/s, elapsed:  5s, eta: 36s
#> 
#> STOMP [===>--------------------------]  12% at 2656 it/s, elapsed:  5s, eta: 35s
#> 
#> STOMP [===>--------------------------]  12% at 2659 it/s, elapsed:  5s, eta: 35s
#> 
#> STOMP [===>--------------------------]  13% at 2659 it/s, elapsed:  5s, eta: 35s
#> 
#> STOMP [===>--------------------------]  13% at 2662 it/s, elapsed:  5s, eta: 35s
#> 
#> STOMP [===>--------------------------]  13% at 2665 it/s, elapsed:  5s, eta: 35s
#> 
#> STOMP [===>--------------------------]  13% at 2669 it/s, elapsed:  5s, eta: 35s
#> 
#> STOMP [===>--------------------------]  14% at 2671 it/s, elapsed:  5s, eta: 35s
#> 
#> STOMP [===>--------------------------]  14% at 2667 it/s, elapsed:  6s, eta: 35s
#> 
#> STOMP [===>--------------------------]  14% at 2661 it/s, elapsed:  6s, eta: 35s
#> 
#> STOMP [===>--------------------------]  14% at 2657 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [===>--------------------------]  15% at 2655 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [===>--------------------------]  15% at 2658 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [====>-------------------------]  15% at 2661 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [====>-------------------------]  15% at 2663 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [====>-------------------------]  16% at 2666 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [====>-------------------------]  16% at 2670 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [====>-------------------------]  16% at 2668 it/s, elapsed:  6s, eta: 34s
#> 
#> STOMP [====>-------------------------]  16% at 2671 it/s, elapsed:  7s, eta: 34s
#> 
#> STOMP [====>-------------------------]  17% at 2672 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [====>-------------------------]  17% at 2670 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [====>-------------------------]  17% at 2661 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [====>-------------------------]  17% at 2660 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [====>-------------------------]  18% at 2653 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [====>-------------------------]  18% at 2655 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [====>-------------------------]  18% at 2659 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [=====>------------------------]  19% at 2659 it/s, elapsed:  7s, eta: 33s
#> 
#> STOMP [=====>------------------------]  19% at 2661 it/s, elapsed:  8s, eta: 33s
#> 
#> STOMP [=====>------------------------]  19% at 2660 it/s, elapsed:  8s, eta: 33s
#> 
#> STOMP [=====>------------------------]  19% at 2659 it/s, elapsed:  8s, eta: 32s
#> 
#> STOMP [=====>------------------------]  20% at 2653 it/s, elapsed:  8s, eta: 32s
#> 
#> STOMP [=====>------------------------]  20% at 2654 it/s, elapsed:  8s, eta: 32s
#> 
#> STOMP [=====>------------------------]  20% at 2650 it/s, elapsed:  8s, eta: 32s
#> 
#> STOMP [=====>------------------------]  20% at 2645 it/s, elapsed:  8s, eta: 32s
#> 
#> STOMP [=====>------------------------]  21% at 2636 it/s, elapsed:  8s, eta: 32s
#> 
#> STOMP [=====>------------------------]  21% at 2629 it/s, elapsed:  8s, eta: 32s
#> 
#> STOMP [=====>------------------------]  21% at 2628 it/s, elapsed:  9s, eta: 32s
#> 
#> STOMP [=====>------------------------]  21% at 2626 it/s, elapsed:  9s, eta: 32s
#> 
#> STOMP [=====>------------------------]  22% at 2624 it/s, elapsed:  9s, eta: 32s
#> 
#> STOMP [======>-----------------------]  22% at 2625 it/s, elapsed:  9s, eta: 32s
#> 
#> STOMP [======>-----------------------]  22% at 2626 it/s, elapsed:  9s, eta: 32s
#> 
#> STOMP [======>-----------------------]  22% at 2630 it/s, elapsed:  9s, eta: 32s
#> 
#> STOMP [======>-----------------------]  23% at 2631 it/s, elapsed:  9s, eta: 31s
#> 
#> STOMP [======>-----------------------]  23% at 2630 it/s, elapsed:  9s, eta: 31s
#> 
#> STOMP [======>-----------------------]  23% at 2632 it/s, elapsed:  9s, eta: 31s
#> 
#> STOMP [======>-----------------------]  23% at 2635 it/s, elapsed:  9s, eta: 31s
#> 
#> STOMP [======>-----------------------]  24% at 2636 it/s, elapsed: 10s, eta: 31s
#> 
#> STOMP [======>-----------------------]  24% at 2635 it/s, elapsed: 10s, eta: 31s
#> 
#> STOMP [======>-----------------------]  24% at 2636 it/s, elapsed: 10s, eta: 31s
#> 
#> STOMP [======>-----------------------]  24% at 2639 it/s, elapsed: 10s, eta: 31s
#> 
#> STOMP [======>-----------------------]  25% at 2640 it/s, elapsed: 10s, eta: 31s
#> 
#> STOMP [======>-----------------------]  25% at 2639 it/s, elapsed: 10s, eta: 30s
#> 
#> STOMP [=======>----------------------]  25% at 2641 it/s, elapsed: 10s, eta: 30s
#> 
#> STOMP [=======>----------------------]  25% at 2635 it/s, elapsed: 10s, eta: 30s
#> 
#> STOMP [=======>----------------------]  26% at 2638 it/s, elapsed: 10s, eta: 30s
#> 
#> STOMP [=======>----------------------]  26% at 2640 it/s, elapsed: 10s, eta: 30s
#> 
#> STOMP [=======>----------------------]  26% at 2641 it/s, elapsed: 11s, eta: 30s
#> 
#> STOMP [=======>----------------------]  26% at 2639 it/s, elapsed: 11s, eta: 30s
#> 
#> STOMP [=======>----------------------]  27% at 2640 it/s, elapsed: 11s, eta: 30s
#> 
#> STOMP [=======>----------------------]  27% at 2641 it/s, elapsed: 11s, eta: 30s
#> 
#> STOMP [=======>----------------------]  27% at 2642 it/s, elapsed: 11s, eta: 30s
#> 
#> STOMP [=======>----------------------]  27% at 2642 it/s, elapsed: 11s, eta: 29s
#> 
#> STOMP [=======>----------------------]  28% at 2642 it/s, elapsed: 11s, eta: 29s
#> 
#> STOMP [=======>----------------------]  28% at 2643 it/s, elapsed: 11s, eta: 29s
#> 
#> STOMP [=======>----------------------]  28% at 2644 it/s, elapsed: 11s, eta: 29s
#> 
#> STOMP [=======>----------------------]  28% at 2645 it/s, elapsed: 11s, eta: 29s
#> 
#> STOMP [========>---------------------]  29% at 2646 it/s, elapsed: 12s, eta: 29s
#> 
#> STOMP [========>---------------------]  29% at 2648 it/s, elapsed: 12s, eta: 29s
#> 
#> STOMP [========>---------------------]  29% at 2646 it/s, elapsed: 12s, eta: 29s
#> 
#> STOMP [========>---------------------]  29% at 2641 it/s, elapsed: 12s, eta: 29s
#> 
#> STOMP [========>---------------------]  30% at 2615 it/s, elapsed: 12s, eta: 29s
#> 
#> STOMP [========>---------------------]  30% at 2593 it/s, elapsed: 12s, eta: 29s
#> 
#> STOMP [========>---------------------]  30% at 2585 it/s, elapsed: 12s, eta: 29s
#> 
#> STOMP [========>---------------------]  30% at 2581 it/s, elapsed: 13s, eta: 29s
#> 
#> STOMP [========>---------------------]  31% at 2575 it/s, elapsed: 13s, eta: 29s
#> 
#> STOMP [========>---------------------]  31% at 2572 it/s, elapsed: 13s, eta: 29s
#> 
#> STOMP [========>---------------------]  31% at 2569 it/s, elapsed: 13s, eta: 29s
#> 
#> STOMP [========>---------------------]  32% at 2565 it/s, elapsed: 13s, eta: 29s
#> 
#> STOMP [=========>--------------------]  32% at 2563 it/s, elapsed: 13s, eta: 28s
#> 
#> STOMP [=========>--------------------]  32% at 2558 it/s, elapsed: 13s, eta: 28s
#> 
#> STOMP [=========>--------------------]  32% at 2552 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  33% at 2548 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  33% at 2545 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  33% at 2546 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  33% at 2549 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  34% at 2551 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  34% at 2554 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  34% at 2553 it/s, elapsed: 14s, eta: 28s
#> 
#> STOMP [=========>--------------------]  35% at 2554 it/s, elapsed: 14s, eta: 27s
#> 
#> STOMP [=========>--------------------]  35% at 2554 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  35% at 2555 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  35% at 2548 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  36% at 2538 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  36% at 2536 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  36% at 2535 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  36% at 2534 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  37% at 2532 it/s, elapsed: 15s, eta: 27s
#> 
#> STOMP [==========>-------------------]  37% at 2532 it/s, elapsed: 16s, eta: 27s
#> 
#> STOMP [==========>-------------------]  37% at 2531 it/s, elapsed: 16s, eta: 27s
#> 
#> STOMP [==========>-------------------]  37% at 2533 it/s, elapsed: 16s, eta: 26s
#> 
#> STOMP [==========>-------------------]  38% at 2535 it/s, elapsed: 16s, eta: 26s
#> 
#> STOMP [==========>-------------------]  38% at 2536 it/s, elapsed: 16s, eta: 26s
#> 
#> STOMP [==========>-------------------]  38% at 2532 it/s, elapsed: 16s, eta: 26s
#> 
#> STOMP [===========>------------------]  38% at 2529 it/s, elapsed: 16s, eta: 26s
#> 
#> STOMP [===========>------------------]  39% at 2530 it/s, elapsed: 16s, eta: 26s
#> 
#> STOMP [===========>------------------]  39% at 2532 it/s, elapsed: 16s, eta: 26s
#> 
#> STOMP [===========>------------------]  39% at 2533 it/s, elapsed: 17s, eta: 26s
#> 
#> STOMP [===========>------------------]  40% at 2532 it/s, elapsed: 17s, eta: 26s
#> 
#> STOMP [===========>------------------]  40% at 2528 it/s, elapsed: 17s, eta: 25s
#> 
#> STOMP [===========>------------------]  40% at 2527 it/s, elapsed: 17s, eta: 25s
#> 
#> STOMP [===========>------------------]  40% at 2526 it/s, elapsed: 17s, eta: 25s
#> 
#> STOMP [===========>------------------]  41% at 2526 it/s, elapsed: 17s, eta: 25s
#> 
#> STOMP [===========>------------------]  41% at 2525 it/s, elapsed: 17s, eta: 25s
#> 
#> STOMP [===========>------------------]  41% at 2526 it/s, elapsed: 18s, eta: 25s
#> 
#> STOMP [===========>------------------]  42% at 2520 it/s, elapsed: 18s, eta: 25s
#> 
#> STOMP [============>-----------------]  42% at 2515 it/s, elapsed: 18s, eta: 25s
#> 
#> STOMP [============>-----------------]  42% at 2513 it/s, elapsed: 18s, eta: 25s
#> 
#> STOMP [============>-----------------]  43% at 2514 it/s, elapsed: 18s, eta: 24s
#> 
#> STOMP [============>-----------------]  43% at 2512 it/s, elapsed: 18s, eta: 24s
#> 
#> STOMP [============>-----------------]  43% at 2508 it/s, elapsed: 18s, eta: 24s
#> 
#> STOMP [============>-----------------]  43% at 2509 it/s, elapsed: 18s, eta: 24s
#> 
#> STOMP [============>-----------------]  44% at 2510 it/s, elapsed: 19s, eta: 24s
#> 
#> STOMP [============>-----------------]  44% at 2511 it/s, elapsed: 19s, eta: 24s
#> 
#> STOMP [============>-----------------]  45% at 2513 it/s, elapsed: 19s, eta: 24s
#> 
#> STOMP [============>-----------------]  45% at 2514 it/s, elapsed: 19s, eta: 23s
#> 
#> STOMP [=============>----------------]  45% at 2514 it/s, elapsed: 19s, eta: 23s
#> 
#> STOMP [=============>----------------]  45% at 2513 it/s, elapsed: 19s, eta: 23s
#> 
#> STOMP [=============>----------------]  46% at 2514 it/s, elapsed: 19s, eta: 23s
#> 
#> STOMP [=============>----------------]  46% at 2514 it/s, elapsed: 20s, eta: 23s
#> 
#> STOMP [=============>----------------]  46% at 2515 it/s, elapsed: 20s, eta: 23s
#> 
#> STOMP [=============>----------------]  47% at 2515 it/s, elapsed: 20s, eta: 23s
#> 
#> STOMP [=============>----------------]  47% at 2514 it/s, elapsed: 20s, eta: 22s
#> 
#> STOMP [=============>----------------]  47% at 2515 it/s, elapsed: 20s, eta: 22s
#> 
#> STOMP [=============>----------------]  48% at 2516 it/s, elapsed: 20s, eta: 22s
#> 
#> STOMP [=============>----------------]  48% at 2518 it/s, elapsed: 20s, eta: 22s
#> 
#> STOMP [==============>---------------]  48% at 2519 it/s, elapsed: 21s, eta: 22s
#> 
#> STOMP [==============>---------------]  49% at 2520 it/s, elapsed: 21s, eta: 22s
#> 
#> STOMP [==============>---------------]  49% at 2519 it/s, elapsed: 21s, eta: 22s
#> 
#> STOMP [==============>---------------]  49% at 2519 it/s, elapsed: 21s, eta: 21s
#> 
#> STOMP [==============>---------------]  50% at 2519 it/s, elapsed: 21s, eta: 21s
#> 
#> STOMP [==============>---------------]  50% at 2518 it/s, elapsed: 21s, eta: 21s
#> 
#> STOMP [==============>---------------]  50% at 2517 it/s, elapsed: 21s, eta: 21s
#> 
#> STOMP [==============>---------------]  50% at 2516 it/s, elapsed: 21s, eta: 21s
#> 
#> STOMP [==============>---------------]  51% at 2517 it/s, elapsed: 22s, eta: 21s
#> 
#> STOMP [==============>---------------]  51% at 2518 it/s, elapsed: 22s, eta: 21s
#> 
#> STOMP [==============>---------------]  52% at 2518 it/s, elapsed: 22s, eta: 21s
#> 
#> STOMP [===============>--------------]  52% at 2518 it/s, elapsed: 22s, eta: 20s
#> 
#> STOMP [===============>--------------]  52% at 2519 it/s, elapsed: 22s, eta: 20s
#> 
#> STOMP [===============>--------------]  53% at 2520 it/s, elapsed: 22s, eta: 20s
#> 
#> STOMP [===============>--------------]  53% at 2521 it/s, elapsed: 22s, eta: 20s
#> 
#> STOMP [===============>--------------]  53% at 2522 it/s, elapsed: 23s, eta: 20s
#> 
#> STOMP [===============>--------------]  53% at 2523 it/s, elapsed: 23s, eta: 20s
#> 
#> STOMP [===============>--------------]  54% at 2524 it/s, elapsed: 23s, eta: 20s
#> 
#> STOMP [===============>--------------]  54% at 2525 it/s, elapsed: 23s, eta: 20s
#> 
#> STOMP [===============>--------------]  54% at 2524 it/s, elapsed: 23s, eta: 19s
#> 
#> STOMP [===============>--------------]  54% at 2516 it/s, elapsed: 23s, eta: 19s
#> 
#> STOMP [===============>--------------]  55% at 2503 it/s, elapsed: 23s, eta: 19s
#> 
#> STOMP [===============>--------------]  55% at 2482 it/s, elapsed: 24s, eta: 19s
#> 
#> STOMP [================>-------------]  55% at 2457 it/s, elapsed: 24s, eta: 20s
#> 
#> STOMP [================>-------------]  55% at 2456 it/s, elapsed: 24s, eta: 19s
#> 
#> STOMP [================>-------------]  56% at 2455 it/s, elapsed: 24s, eta: 19s
#> 
#> STOMP [================>-------------]  56% at 2456 it/s, elapsed: 24s, eta: 19s
#> 
#> STOMP [================>-------------]  56% at 2455 it/s, elapsed: 25s, eta: 19s
#> 
#> STOMP [================>-------------]  57% at 2454 it/s, elapsed: 25s, eta: 19s
#> 
#> STOMP [================>-------------]  57% at 2450 it/s, elapsed: 25s, eta: 19s
#> 
#> STOMP [================>-------------]  57% at 2438 it/s, elapsed: 25s, eta: 19s
#> 
#> STOMP [================>-------------]  58% at 2432 it/s, elapsed: 25s, eta: 19s
#> 
#> STOMP [================>-------------]  58% at 2430 it/s, elapsed: 25s, eta: 19s
#> 
#> STOMP [================>-------------]  58% at 2430 it/s, elapsed: 26s, eta: 18s
#> 
#> STOMP [=================>------------]  58% at 2427 it/s, elapsed: 26s, eta: 18s
#> 
#> STOMP [=================>------------]  59% at 2423 it/s, elapsed: 26s, eta: 18s
#> 
#> STOMP [=================>------------]  59% at 2415 it/s, elapsed: 26s, eta: 18s
#> 
#> STOMP [=================>------------]  59% at 2411 it/s, elapsed: 26s, eta: 18s
#> 
#> STOMP [=================>------------]  60% at 2410 it/s, elapsed: 26s, eta: 18s
#> 
#> STOMP [=================>------------]  60% at 2411 it/s, elapsed: 27s, eta: 18s
#> 
#> STOMP [=================>------------]  60% at 2412 it/s, elapsed: 27s, eta: 18s
#> 
#> STOMP [=================>------------]  61% at 2412 it/s, elapsed: 27s, eta: 17s
#> 
#> STOMP [=================>------------]  61% at 2414 it/s, elapsed: 27s, eta: 17s
#> 
#> STOMP [=================>------------]  62% at 2413 it/s, elapsed: 27s, eta: 17s
#> 
#> STOMP [==================>-----------]  62% at 2413 it/s, elapsed: 27s, eta: 17s
#> 
#> STOMP [==================>-----------]  62% at 2413 it/s, elapsed: 28s, eta: 17s
#> 
#> STOMP [==================>-----------]  63% at 2413 it/s, elapsed: 28s, eta: 17s
#> 
#> STOMP [==================>-----------]  63% at 2414 it/s, elapsed: 28s, eta: 16s
#> 
#> STOMP [==================>-----------]  63% at 2415 it/s, elapsed: 28s, eta: 16s
#> 
#> STOMP [==================>-----------]  64% at 2416 it/s, elapsed: 28s, eta: 16s
#> 
#> STOMP [==================>-----------]  64% at 2417 it/s, elapsed: 28s, eta: 16s
#> 
#> STOMP [==================>-----------]  65% at 2418 it/s, elapsed: 29s, eta: 16s
#> 
#> STOMP [==================>-----------]  65% at 2419 it/s, elapsed: 29s, eta: 16s
#> 
#> STOMP [===================>----------]  65% at 2419 it/s, elapsed: 29s, eta: 15s
#> 
#> STOMP [===================>----------]  66% at 2419 it/s, elapsed: 29s, eta: 15s
#> 
#> STOMP [===================>----------]  66% at 2420 it/s, elapsed: 29s, eta: 15s
#> 
#> STOMP [===================>----------]  66% at 2421 it/s, elapsed: 29s, eta: 15s
#> 
#> STOMP [===================>----------]  66% at 2420 it/s, elapsed: 29s, eta: 15s
#> 
#> STOMP [===================>----------]  67% at 2414 it/s, elapsed: 30s, eta: 15s
#> 
#> STOMP [===================>----------]  67% at 2413 it/s, elapsed: 30s, eta: 15s
#> 
#> STOMP [===================>----------]  67% at 2413 it/s, elapsed: 30s, eta: 14s
#> 
#> STOMP [===================>----------]  68% at 2413 it/s, elapsed: 30s, eta: 14s
#> 
#> STOMP [===================>----------]  68% at 2414 it/s, elapsed: 30s, eta: 14s
#> 
#> STOMP [====================>---------]  68% at 2415 it/s, elapsed: 30s, eta: 14s
#> 
#> STOMP [====================>---------]  69% at 2416 it/s, elapsed: 30s, eta: 14s
#> 
#> STOMP [====================>---------]  69% at 2417 it/s, elapsed: 30s, eta: 14s
#> 
#> STOMP [====================>---------]  69% at 2417 it/s, elapsed: 31s, eta: 14s
#> 
#> STOMP [====================>---------]  69% at 2419 it/s, elapsed: 31s, eta: 14s
#> 
#> STOMP [====================>---------]  70% at 2420 it/s, elapsed: 31s, eta: 13s
#> 
#> STOMP [====================>---------]  70% at 2422 it/s, elapsed: 31s, eta: 13s
#> 
#> STOMP [====================>---------]  70% at 2423 it/s, elapsed: 31s, eta: 13s
#> 
#> STOMP [====================>---------]  71% at 2425 it/s, elapsed: 31s, eta: 13s
#> 
#> STOMP [====================>---------]  71% at 2426 it/s, elapsed: 31s, eta: 13s
#> 
#> STOMP [====================>---------]  71% at 2427 it/s, elapsed: 31s, eta: 13s
#> 
#> STOMP [=====================>--------]  72% at 2428 it/s, elapsed: 32s, eta: 12s
#> 
#> STOMP [=====================>--------]  72% at 2429 it/s, elapsed: 32s, eta: 12s
#> 
#> STOMP [=====================>--------]  73% at 2429 it/s, elapsed: 32s, eta: 12s
#> 
#> STOMP [=====================>--------]  73% at 2428 it/s, elapsed: 32s, eta: 12s
#> 
#> STOMP [=====================>--------]  73% at 2429 it/s, elapsed: 32s, eta: 12s
#> 
#> STOMP [=====================>--------]  73% at 2430 it/s, elapsed: 32s, eta: 12s
#> 
#> STOMP [=====================>--------]  74% at 2430 it/s, elapsed: 32s, eta: 12s
#> 
#> STOMP [=====================>--------]  74% at 2430 it/s, elapsed: 33s, eta: 11s
#> 
#> STOMP [=====================>--------]  74% at 2431 it/s, elapsed: 33s, eta: 11s
#> 
#> STOMP [=====================>--------]  75% at 2432 it/s, elapsed: 33s, eta: 11s
#> 
#> STOMP [======================>-------]  75% at 2433 it/s, elapsed: 33s, eta: 11s
#> 
#> STOMP [======================>-------]  75% at 2434 it/s, elapsed: 33s, eta: 11s
#> 
#> STOMP [======================>-------]  76% at 2434 it/s, elapsed: 33s, eta: 11s
#> 
#> STOMP [======================>-------]  76% at 2435 it/s, elapsed: 33s, eta: 10s
#> 
#> STOMP [======================>-------]  76% at 2435 it/s, elapsed: 34s, eta: 10s
#> 
#> STOMP [======================>-------]  77% at 2435 it/s, elapsed: 34s, eta: 10s
#> 
#> STOMP [======================>-------]  77% at 2436 it/s, elapsed: 34s, eta: 10s
#> 
#> STOMP [======================>-------]  77% at 2437 it/s, elapsed: 34s, eta: 10s
#> 
#> STOMP [======================>-------]  78% at 2439 it/s, elapsed: 34s, eta: 10s
#> 
#> STOMP [======================>-------]  78% at 2440 it/s, elapsed: 34s, eta: 10s
#> 
#> STOMP [======================>-------]  78% at 2441 it/s, elapsed: 34s, eta: 10s
#> 
#> STOMP [=======================>------]  78% at 2442 it/s, elapsed: 34s, eta:  9s
#> 
#> STOMP [=======================>------]  79% at 2442 it/s, elapsed: 34s, eta:  9s
#> 
#> STOMP [=======================>------]  79% at 2443 it/s, elapsed: 35s, eta:  9s
#> 
#> STOMP [=======================>------]  79% at 2444 it/s, elapsed: 35s, eta:  9s
#> 
#> STOMP [=======================>------]  79% at 2446 it/s, elapsed: 35s, eta:  9s
#> 
#> STOMP [=======================>------]  80% at 2447 it/s, elapsed: 35s, eta:  9s
#> 
#> STOMP [=======================>------]  80% at 2448 it/s, elapsed: 35s, eta:  9s
#> 
#> STOMP [=======================>------]  80% at 2449 it/s, elapsed: 35s, eta:  9s
#> 
#> STOMP [=======================>------]  81% at 2450 it/s, elapsed: 35s, eta:  8s
#> 
#> STOMP [=======================>------]  81% at 2451 it/s, elapsed: 35s, eta:  8s
#> 
#> STOMP [=======================>------]  81% at 2452 it/s, elapsed: 35s, eta:  8s
#> 
#> STOMP [=======================>------]  81% at 2453 it/s, elapsed: 36s, eta:  8s
#> 
#> STOMP [========================>-----]  82% at 2454 it/s, elapsed: 36s, eta:  8s
#> 
#> STOMP [========================>-----]  82% at 2453 it/s, elapsed: 36s, eta:  8s
#> 
#> STOMP [========================>-----]  82% at 2449 it/s, elapsed: 36s, eta:  8s
#> 
#> STOMP [========================>-----]  82% at 2445 it/s, elapsed: 36s, eta:  8s
#> 
#> STOMP [========================>-----]  83% at 2446 it/s, elapsed: 36s, eta:  8s
#> 
#> STOMP [========================>-----]  83% at 2446 it/s, elapsed: 36s, eta:  7s
#> 
#> STOMP [========================>-----]  83% at 2447 it/s, elapsed: 36s, eta:  7s
#> 
#> STOMP [========================>-----]  83% at 2448 it/s, elapsed: 36s, eta:  7s
#> 
#> STOMP [========================>-----]  84% at 2448 it/s, elapsed: 37s, eta:  7s
#> 
#> STOMP [========================>-----]  84% at 2447 it/s, elapsed: 37s, eta:  7s
#> 
#> STOMP [========================>-----]  85% at 2444 it/s, elapsed: 37s, eta:  7s
#> 
#> STOMP [========================>-----]  85% at 2442 it/s, elapsed: 37s, eta:  7s
#> 
#> STOMP [=========================>----]  85% at 2442 it/s, elapsed: 37s, eta:  6s
#> 
#> STOMP [=========================>----]  85% at 2441 it/s, elapsed: 37s, eta:  6s
#> 
#> STOMP [=========================>----]  86% at 2440 it/s, elapsed: 38s, eta:  6s
#> 
#> STOMP [=========================>----]  86% at 2438 it/s, elapsed: 38s, eta:  6s
#> 
#> STOMP [=========================>----]  86% at 2439 it/s, elapsed: 38s, eta:  6s
#> 
#> STOMP [=========================>----]  87% at 2440 it/s, elapsed: 38s, eta:  6s
#> 
#> STOMP [=========================>----]  87% at 2439 it/s, elapsed: 38s, eta:  6s
#> 
#> STOMP [=========================>----]  87% at 2440 it/s, elapsed: 38s, eta:  6s
#> 
#> STOMP [=========================>----]  87% at 2439 it/s, elapsed: 38s, eta:  5s
#> 
#> STOMP [=========================>----]  88% at 2438 it/s, elapsed: 38s, eta:  5s
#> 
#> STOMP [=========================>----]  88% at 2439 it/s, elapsed: 39s, eta:  5s
#> 
#> STOMP [==========================>---]  88% at 2439 it/s, elapsed: 39s, eta:  5s
#> 
#> STOMP [==========================>---]  89% at 2440 it/s, elapsed: 39s, eta:  5s
#> 
#> STOMP [==========================>---]  89% at 2441 it/s, elapsed: 39s, eta:  5s
#> 
#> STOMP [==========================>---]  90% at 2441 it/s, elapsed: 39s, eta:  5s
#> 
#> STOMP [==========================>---]  90% at 2441 it/s, elapsed: 39s, eta:  4s
#> 
#> STOMP [==========================>---]  90% at 2442 it/s, elapsed: 40s, eta:  4s
#> 
#> STOMP [==========================>---]  90% at 2443 it/s, elapsed: 40s, eta:  4s
#> 
#> STOMP [==========================>---]  91% at 2443 it/s, elapsed: 40s, eta:  4s
#> 
#> STOMP [==========================>---]  91% at 2444 it/s, elapsed: 40s, eta:  4s
#> 
#> STOMP [==========================>---]  91% at 2445 it/s, elapsed: 40s, eta:  4s
#> 
#> STOMP [===========================>--]  92% at 2446 it/s, elapsed: 40s, eta:  4s
#> 
#> STOMP [===========================>--]  92% at 2447 it/s, elapsed: 40s, eta:  4s
#> 
#> STOMP [===========================>--]  92% at 2448 it/s, elapsed: 40s, eta:  3s
#> 
#> STOMP [===========================>--]  92% at 2447 it/s, elapsed: 40s, eta:  3s
#> 
#> STOMP [===========================>--]  93% at 2447 it/s, elapsed: 41s, eta:  3s
#> 
#> STOMP [===========================>--]  93% at 2448 it/s, elapsed: 41s, eta:  3s
#> 
#> STOMP [===========================>--]  93% at 2449 it/s, elapsed: 41s, eta:  3s
#> 
#> STOMP [===========================>--]  94% at 2450 it/s, elapsed: 41s, eta:  3s
#> 
#> STOMP [===========================>--]  94% at 2451 it/s, elapsed: 41s, eta:  3s
#> 
#> STOMP [===========================>--]  94% at 2452 it/s, elapsed: 41s, eta:  2s
#> 
#> STOMP [===========================>--]  95% at 2451 it/s, elapsed: 41s, eta:  2s
#> 
#> STOMP [===========================>--]  95% at 2452 it/s, elapsed: 41s, eta:  2s
#> 
#> STOMP [============================>-]  95% at 2453 it/s, elapsed: 42s, eta:  2s
#> 
#> STOMP [============================>-]  96% at 2453 it/s, elapsed: 42s, eta:  2s
#> 
#> STOMP [============================>-]  97% at 2452 it/s, elapsed: 42s, eta:  1s
#> 
#> STOMP [============================>-]  97% at 2453 it/s, elapsed: 42s, eta:  1s
#> 
#> STOMP [============================>-]  98% at 2454 it/s, elapsed: 43s, eta:  1s
#> 
#> STOMP [============================>-]  98% at 2455 it/s, elapsed: 43s, eta:  1s
#> 
#> STOMP [=============================>]  98% at 2454 it/s, elapsed: 43s, eta:  1s
#> 
#> STOMP [=============================>]  99% at 2455 it/s, elapsed: 43s, eta:  1s
#> 
#> STOMP [=============================>]  99% at 2456 it/s, elapsed: 43s, eta:  0s
#> 
#> STOMP [=============================>]  99% at 2455 it/s, elapsed: 43s, eta:  0s
#> 
#> STOMP [=============================>] 100% at 2455 it/s, elapsed: 43s, eta:  0s
#> 
#> STOMP [==============================] 100% at 2454 it/s, elapsed: 44s, eta:  0s
#> 
#> Error in {    work_len <- length(idx_work[[i]])    pro_muls <- matrix(Inf, matrix_profile_size, 1)    pro_idxs <- matrix(-Inf, matrix_profile_size, 1)    if (join) {        pro_muls_right <- pro_muls_left <- NULL        pro_idxs_right <- pro_idxs_left <- NULL    }    else {        pro_muls_right <- pro_muls_left <- pro_muls        pro_idxs_right <- pro_idxs_left <- pro_idxs    }    dist_pro <- matrix(0, matrix_profile_size, 1)    last_product <- matrix(0, matrix_profile_size, 1)    drop_value <- matrix(0, 1, 1)    for (j in 1:work_len) {        idx_st <- idx_work[[i]][1]        idx_ed <- idx_work[[i]][work_len]        idx <- idx_work[[i]][j]        query_window <- as.matrix(query[idx:(idx + window_size -             1), 1])        if (j == 1) {            nni <- dist_profile(data, query, nn, index = idx)            dist_pro[, 1] <- nni$distance_profile            last_product[, 1] <- nni$last_product        }        else {            last_product[2:(data_size - window_size + 1), 1] <- last_product[1:(data_size -                 window_size), 1] - data[1:(data_size - window_size),                 1] * drop_value + data[(window_size + 1):data_size,                 1] * query_window[window_size, 1]            last_product[1, 1] <- first_product[idx, 1]            dist_pro <- 2 * (window_size - (last_product - window_size *                 nni$par$data_mean * nni$par$query_mean[idx])/(nni$par$data_sd *                 nni$par$query_sd[idx]))        }        dist_pro[dist_pro < 0] <- 0        dist_pro <- sqrt(dist_pro)        drop_value <- query_window[1, 1]        if (exclusion_zone > 0) {            exc_st <- max(1, idx - exclusion_zone)            exc_ed <- min(matrix_profile_size, idx + exclusion_zone)            dist_pro[exc_st:exc_ed, 1] <- Inf        }        dist_pro[nni$par$data_sd < vars()$eps] <- Inf        if (skip_location[idx] || any(nni$par$query_sd[idx] <             vars()$eps)) {            dist_pro[] <- Inf        }        dist_pro[skip_location] <- Inf        if (!join) {            ind <- (dist_pro[idx:matrix_profile_size] < pro_muls_left[idx:matrix_profile_size])            ind <- c(rep(FALSE, (idx - 1)), ind)            pro_muls_left[ind] <- dist_pro[ind]            pro_idxs_left[which(ind)] <- idx            ind <- (dist_pro[1:idx] < pro_muls_right[1:idx])            ind <- c(ind, rep(FALSE, matrix_profile_size - idx))            pro_muls_right[ind] <- dist_pro[ind]            pro_idxs_right[which(ind)] <- idx        }        ind <- (dist_pro < pro_muls)        pro_muls[ind] <- dist_pro[ind]        pro_idxs[which(ind)] <- idx    }    res <- list(pro_muls = pro_muls, pro_idxs = pro_idxs, pro_muls_left = pro_muls_left,         pro_idxs_left = pro_idxs_left, pro_muls_right = pro_muls_right,         pro_idxs_right = pro_idxs_right)    res}: task 1 failed - "could not find function "mass_v3""
mps <- salient_subsequences(mp, data, n_bits = c(4, 6, 8), verbose = 2)
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#> Warning: number of rows of result is not a multiple of vector length (arg 2)
#> Finished in 0.05 secs
# }