Anytime univariate STAMP algorithm Parallel version

Computes the best so far Matrix Profile and Profile Index for Univariate Time Series.

stamp_par(
  ...,
  window_size,
  exclusion_zone = getOption("tsmp.exclusion_zone", 1/2),
  verbose = getOption("tsmp.verbose", 2),
  s_size = Inf,
  n_workers = 2,
  weight = NULL
)

stamp(
  ...,
  window_size,
  exclusion_zone = getOption("tsmp.exclusion_zone", 1/2),
  verbose = getOption("tsmp.verbose", 2),
  s_size = Inf,
  weight = NULL
)

Arguments

...

a matrix or a vector. If a second time series is supplied it will be a join matrix profile.

window_size

an int. Size of the sliding window.

exclusion_zone

a numeric. Size of the exclusion zone, based on window size (default is 1/2). See details.

verbose

an int. See details. (Default is 2).

s_size

a numeric. for anytime algorithm, represents the size (in observations) the random calculation will occur (default is Inf).

n_workers

an int. Number of workers for parallel. (Default is 2).

weight

a vector of numeric or NULL with the same length of the window_size. This is a MASS extension to weight the query.

Value

Returns a MatrixProfile object, a list with the matrix profile mp, profile index pi

left and right matrix profile lmp, rmp and profile index lpi, rpi, window size w and exclusion zone ez.

Details

The Matrix Profile, has the potential to revolutionize time series data mining because of its generality, versatility, simplicity and scalability. In particular it has implications for time series motif discovery, time series joins, shapelet discovery (classification), density estimation, semantic segmentation, visualization, rule discovery, clustering etc. The anytime STAMP computes the Matrix Profile and Profile Index in such manner that it can be stopped before its complete calculation and return the best so far results allowing ultra-fast approximate solutions. verbose changes how much information is printed by this function; 0 means nothing, 1 means text, 2 adds the progress bar, 3 adds the finish sound. exclusion_zone is used to avoid trivial matches; if a query data is provided (join similarity), this parameter is ignored.

Functions

stamp_par(): Parallel version.

stamp(): Single thread version.

References

Yeh CCM, Zhu Y, Ulanova L, Begum N, Ding Y, Dau HA, et al. Matrix profile I: All pairs similarity joins for time series: A unifying view that includes motifs, discords and shapelets. Proc - IEEE Int Conf Data Mining, ICDM. 2017;1317-22.

Zhu Y, Imamura M, Nikovski D, Keogh E. Matrix Profile VII: Time Series Chains: A New Primitive for Time Series Data Mining. Knowl Inf Syst. 2018 Jun 2;1-27.

Website: http://www.cs.ucr.edu/~eamonn/MatrixProfile.html

See also

Examples

mp <- stamp(mp_toy_data$data[1:200, 1], window_size = 30, verbose = 0)
# \donttest{

#' # using threads
mp <- stamp_par(mp_toy_data$data[1:200, 1], window_size = 30, verbose = 0)
#> Error in {    res <- NULL    index <- k * cols + j    if (index <= ssize) {        i <- order[index]        if (is.null(weight)) {            nn <- dist_profile(data, query, pre, index = i)        }        else {            nn <- dist_profile(data, query, pre, index = i, method = "weighted")        }        distance_profile <- sqrt(nn$distance_profile)        if (exclusion_zone > 0) {            exc_st <- max(1, i - exclusion_zone)            exc_ed <- min(matrix_profile_size, i + exclusion_zone)            distance_profile[exc_st:exc_ed] <- Inf        }        distance_profile[pre$data_sd < vars()$eps] <- Inf        if (skip_location[i] || any(pre$query_sd[i] < vars()$eps)) {            distance_profile[] <- Inf        }        distance_profile[skip_location] <- Inf        res <- list(dp = distance_profile, i = i)    }    res}: task 1 failed - "could not find function "mass_v3""

ref_data <- mp_toy_data$data[, 1]
query_data <- mp_toy_data$data[, 2]
# self similarity
mp <- stamp(ref_data, window_size = 30, s_size = round(nrow(ref_data) * 0.1))
#> 
#> STAMP [========>----------------------]  28% at 726 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  28% at 724 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  28% at 723 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 719 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 717 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 716 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 715 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 691 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 689 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 687 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 686 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 685 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 684 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  31% at 683 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  31% at 681 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  31% at 679 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  31% at 678 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  31% at 677 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  31% at 676 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  32% at 675 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  32% at 674 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  32% at 673 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  33% at 672 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  33% at 671 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  33% at 670 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  33% at 669 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  33% at 668 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  34% at 667 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  34% at 666 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  34% at 664 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  35% at 663 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  35% at 662 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 662 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 661 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 659 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 658 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  37% at 657 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  37% at 656 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  37% at 656 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  37% at 641 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  38% at 640 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  39% at 640 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  39% at 636 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  39% at 636 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  39% at 635 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  40% at 635 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  40% at 634 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  40% at 633 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===========>-------------------]  40% at 630 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  40% at 629 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  41% at 629 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  41% at 628 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  42% at 627 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  42% at 626 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  43% at 626 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  43% at 625 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [============>------------------]  43% at 624 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  44% at 624 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  44% at 623 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  44% at 622 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  45% at 622 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  45% at 621 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  45% at 620 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  45% at 610 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  46% at 609 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  46% at 608 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  47% at 608 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  47% at 608 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  47% at 607 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  48% at 606 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  49% at 606 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  50% at 605 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  50% at 605 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  51% at 605 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  51% at 604 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  52% at 604 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  53% at 604 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  53% at 603 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  53% at 603 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  54% at 602 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  54% at 594 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  54% at 593 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  55% at 593 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  56% at 593 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  56% at 592 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  56% at 593 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=================>-------------]  57% at 592 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=================>-------------]  57% at 592 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  57% at 591 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  58% at 591 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  59% at 591 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  60% at 591 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  60% at 591 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  61% at 591 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  62% at 591 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  62% at 584 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  63% at 584 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  63% at 584 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  63% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  64% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  65% at 584 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  65% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  66% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  66% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  67% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  68% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  69% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  69% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  70% at 583 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  70% at 577 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  71% at 577 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  71% at 576 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  72% at 575 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  72% at 574 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  72% at 573 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  73% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  73% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  73% at 573 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  74% at 573 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  74% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  74% at 573 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  74% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  75% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  76% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  76% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  77% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  78% at 572 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  79% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  79% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  80% at 568 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  81% at 568 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  81% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  82% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  82% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  83% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  84% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  85% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  86% at 567 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  86% at 568 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  87% at 568 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  87% at 563 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  88% at 563 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  88% at 562 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  88% at 561 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  89% at 561 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  89% at 560 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  89% at 559 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  89% at 558 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  90% at 558 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  91% at 558 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  91% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  92% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  92% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  93% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  94% at 558 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  94% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  95% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  95% at 553 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  95% at 553 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  96% at 553 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  96% at 552 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  97% at 552 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  98% at 552 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==============================>]  98% at 552 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==============================>]  99% at 552 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==============================>] 100% at 552 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===============================] 100% at 552 it/s, elapsed:  1s, eta:  0s
#> 
#> Finished in 0.95 secs
# join similarity
mp <- stamp(ref_data, query_data, window_size = 30, s_size = round(nrow(query_data) * 0.1))
#> 
#> STAMP [=======>-----------------------]  26% at 674 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=======>-----------------------]  26% at 672 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=======>-----------------------]  26% at 670 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=======>-----------------------]  27% at 665 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=======>-----------------------]  27% at 663 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=======>-----------------------]  27% at 662 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=======>-----------------------]  27% at 660 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  27% at 659 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  28% at 657 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  28% at 655 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  28% at 651 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  28% at 649 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  28% at 647 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 645 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 643 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 641 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 639 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  29% at 637 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 635 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 633 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 631 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 629 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  30% at 627 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [========>----------------------]  31% at 623 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  31% at 621 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  31% at 620 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  32% at 620 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  32% at 618 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  32% at 616 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  32% at 600 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  33% at 600 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  33% at 599 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=========>---------------------]  34% at 599 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  34% at 599 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  35% at 598 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  35% at 599 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 599 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 598 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 599 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  36% at 597 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  37% at 597 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [==========>--------------------]  37% at 596 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  37% at 596 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  38% at 596 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  39% at 596 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  40% at 597 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  40% at 596 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [===========>-------------------]  40% at 595 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  40% at 595 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  41% at 583 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  41% at 584 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  42% at 584 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  42% at 583 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  42% at 584 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  42% at 581 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [============>------------------]  43% at 581 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=============>-----------------]  44% at 581 it/s, elapsed:  0s, eta:  1s
#> 
#> STAMP [=============>-----------------]  44% at 581 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  45% at 581 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  46% at 581 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  46% at 580 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [=============>-----------------]  47% at 580 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  47% at 580 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  48% at 580 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  49% at 580 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  49% at 571 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  49% at 570 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [==============>----------------]  50% at 570 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  50% at 570 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  51% at 570 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  51% at 571 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  51% at 570 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  52% at 570 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  52% at 568 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [===============>---------------]  53% at 568 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  53% at 568 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  54% at 568 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  55% at 568 it/s, elapsed:  0s, eta:  0s
#> 
#> STAMP [================>--------------]  55% at 568 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [================>--------------]  56% at 569 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [================>--------------]  56% at 568 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [================>--------------]  56% at 569 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  57% at 569 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  57% at 568 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  57% at 561 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  58% at 560 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  58% at 558 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  58% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  59% at 557 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  59% at 555 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  59% at 554 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  59% at 553 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=================>-------------]  60% at 551 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  60% at 550 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  60% at 549 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  60% at 548 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  60% at 547 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  60% at 546 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  61% at 545 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  61% at 544 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  61% at 545 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  62% at 545 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==================>------------]  63% at 545 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  63% at 545 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  63% at 546 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  64% at 546 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  65% at 546 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  65% at 547 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  65% at 539 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  66% at 538 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  66% at 537 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===================>-----------]  66% at 536 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  66% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  67% at 536 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  68% at 537 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  69% at 537 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [====================>----------]  69% at 538 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  69% at 538 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  70% at 538 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  71% at 538 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  71% at 539 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  72% at 539 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=====================>---------]  73% at 540 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  73% at 540 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  73% at 539 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  74% at 538 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  74% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  75% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  75% at 534 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [======================>--------]  76% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  76% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  76% at 534 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  77% at 534 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  78% at 534 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  78% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=======================>-------]  79% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  79% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  80% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  80% at 536 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  81% at 536 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  81% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  81% at 536 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  82% at 536 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [========================>------]  82% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  82% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  83% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  83% at 532 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  84% at 532 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=========================>-----]  85% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  86% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  87% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  88% at 534 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==========================>----]  89% at 534 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  89% at 534 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  89% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  90% at 535 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  91% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===========================>---]  92% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  92% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  93% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  94% at 531 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [============================>--]  95% at 532 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  95% at 532 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  96% at 532 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  97% at 532 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  97% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [=============================>-]  98% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==============================>]  98% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==============================>]  99% at 533 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==============================>]  99% at 529 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [==============================>] 100% at 530 it/s, elapsed:  1s, eta:  0s
#> 
#> STAMP [===============================] 100% at 529 it/s, elapsed:  1s, eta:  0s
#> 
#> Finished in 0.99 secs
# }